+++ src/TypeProviderInstantiation.Experimental.fs
@@ -3,7 +3,8 @@
 open ProviderImplementation.ProvidedTypes
 
 type ApiaryProviderArgs = 
-    { ApiName : string }
+    { ApiName : string 
+      SpecialNames : string}
 
 type TypeProviderInstantiation = 
     | Apiary of ApiaryProviderArgs
@@ -13,20 +14,23 @@ type TypeProviderInstantiation =
             match x with
             | Apiary x ->
                 (fun cfg -> new ApiaryProvider(cfg) :> TypeProviderForNamespaces),
-                [| box x.ApiName |] 
+                [| box x.ApiName
+                   box x.SpecialNames |] 
         Debug.generate resolutionFolder runtimeAssembly platform f args
 
     override x.ToString() =
         match x with
         | Apiary x -> ["Apiary"
-                       x.ApiName]
+                       x.ApiName
+                       x.SpecialNames]
         |> String.concat ","
 
     static member Parse (line:string) =
         let args = line.Split [|','|]
         match args.[0] with
         | "Apiary" ->
-            Apiary { ApiName = args.[1] }
+            Apiary { ApiName = args.[1]
+                     SpecialNames = args.[2].Replace(';', ',') }
         | _ -> failwithf "Unknown: %s" args.[0]
 
 open System.Runtime.CompilerServices