+++ RELEASE_NOTES.md
@@ -140,3 +140,9 @@
 
 #### 2.0.0-alpha7 - Unreleased
 * Improved name generation algorithm to cope better with acronymns.
+* Fixed wrong singularization of words ending with 'uses'.
+* Fixed handling of repeated one letter names.
+* Improve HTTP error messages.
+* Support for more api patterns in ApiaryProvider.
+* Tolerate invalid json and missing data in apiary specifications.
+* Added 'SpecialNames' parameter to ApiaryProvider to allow taking concrete examples in the apiary specification and turning them to api parameters.